Delete unused DLS actions, types
