+++ api/ui/constants/common.ts
@@ -1,7 +1,5 @@
 export {
     DL,
-    PERMISSION,
-    SCOPE_WITH_EXECUTE,
     PRODUCT_NAME,
     DLS_SUBJECT,
     CURRENT_CLOUD_STORE_KEY,
