+++ src/shared/schema/us/types/index.ts
@@ -3,7 +3,6 @@ export * from './common';
 export * from './entries';
 export * from './presets';
 export * from './locks';
-export * from './dls';
 export * from './favorites';
 export * from './editor';
 export * from './state';
