+++ src/shared/types/index.ts
@@ -3,7 +3,6 @@ export * from './configs';
 export * from './connections';
 export * from './dash';
 export * from './widget';
-export * from './dls';
 export * from './dataset';
 export * from './wizard';
 export * from './ql/common';
@@ -24,6 +23,7 @@ export * from './chartkit/dl-chartkit';
 export * from './chartkit/table';
 export * from './config/ql';
 export * from './typed-query-api';
+export * from './permissions';
 
 // TODO: will be removed in CHARTS-10240
 export type GoogleOAuthClient = any;
