+++ src/shared/types/permissions.ts
@@ -1,0 +1,6 @@
+export interface Permissions {
+    read: boolean;
+    edit: boolean;
+    admin: boolean;
+    execute: boolean;
+}
